+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/view/edit/alarm/AlarmEditFragment.kt
@@ -8,7 +8,6 @@ import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager.HORIZONTAL
 import com.bottari.presentation.R
-import com.bottari.presentation.common.ItemSpacingDecoration
 import com.bottari.presentation.common.base.BaseFragment
 import com.bottari.presentation.common.extension.getParcelableCompat
 import com.bottari.presentation.common.extension.getSSAID
@@ -16,6 +15,7 @@ import com.bottari.presentation.common.extension.safeArgument
 import com.bottari.presentation.databinding.FragmentAlarmEditBinding
 import com.bottari.presentation.model.AlarmTypeUiModel
 import com.bottari.presentation.model.AlarmUiModel
+import com.bottari.presentation.view.common.decoration.ItemSpacingDecoration
 import com.bottari.presentation.view.edit.alarm.adapter.DayOfWeekAdapter
 import com.shawnlin.numberpicker.NumberPicker
 import java.time.LocalDate
